status-server: Fix conceptual error in go
There's that thing about only using >! and <! inside go blocks. I'd
thought that everything that is called inside a go block can also use >!
and <!, but of course this is wrong. – >! and <! are special symbols to
be recognised by the go macro, so they have to be syntactically inside
the go block.

Fix my error.